add: AddOwner and use User instead of id - GREEN and REFACTOR

+++ SmartHome/IBusinessLogic/IHomeService.cs
@@ -9,7 +9,7 @@ public interface IHomeService
     List<Home> GetHomesByFilter(Func<Home, bool> filter);
     List<DeviceUnit> GetDevicesFromHome(int homeId);
     List<Member> GetMembersFromHome(long homeId);
-    void AddMemberToHome(int homeId, Member member);
+    void AddMemberToHome(long homeId, Member member);
     Home GetHomeById(long id);
     void PutDevicesInHome(long homeId, List<DeviceUnitDTO> devices);
 }
